+++ client/src/pages/GameEditPage/components/GameUpdateInnerForm.tsx
@@ -6,10 +6,11 @@ import { Button, DatePicker, Form, Input } from 'antd';
 import dayjs from 'dayjs';
 import { ErrorMessage, FieldArray, FormikProps } from 'formik';
 
-import { GameCreateFormValuesType } from '../../GameEditPageTypes';
-import style from '../GameEditForm.module.scss';
+import { UpdateGameFormValuesType } from '../GameEditPageTypes';
 
-export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>) => {
+import style from './GameEditForm.module.scss';
+
+export const UpdateGameInnerForm = (props: FormikProps<UpdateGameFormValuesType>) => {
   const { touched, errors, values, handleBlur } = props;
   const { handleChange, handleSubmit, setFieldValue } = props;
 
@@ -17,7 +18,7 @@ export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>
 
   return (
     <Form className={style.editGameForm} labelCol={{ span: 4 }}>
-      <Form.Item name="basic" required label="Название">
+      <Form.Item required label="Название">
         <div>
           <Input
             onBlur={handleBlur}
@@ -32,14 +33,16 @@ export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>
       </Form.Item>
 
       <Form.Item required label="Дата релиза">
-        <DatePicker
-          onBlur={handleBlur}
-          status={errors.release_date && touched.release_date ? 'error' : ''}
-          allowClear={false}
-          onChange={(__, dateString) => setFieldValue('release_date', dayjs(dateString).format('YYYY-MM-DD'))}
-          value={dayjs(values.release_date)}
-          name="release_date"
-        />
+        <div>
+          <DatePicker
+            onBlur={handleBlur}
+            status={errors.release_date && touched.release_date ? 'error' : ''}
+            allowClear={false}
+            onChange={(__, dateString) => setFieldValue('release_date', dayjs(dateString).format('YYYY-MM-DD'))}
+            value={dayjs(values.release_date)}
+            name="release_date"
+          />
+        </div>
       </Form.Item>
 
       <Form.Item required label="Платформа">
@@ -53,20 +56,6 @@ export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>
             placeholder="Платформа"
           />
           <ErrorMessage name="platform" />
-        </div>
-      </Form.Item>
-
-      <Form.Item required label="slug">
-        <div>
-          <Input
-            onBlur={handleBlur}
-            onChange={handleChange}
-            status={errors.slug && touched.slug ? 'error' : ''}
-            value={values.slug}
-            name="slug"
-            placeholder="slug"
-          />
-          <ErrorMessage name="slug" />
         </div>
       </Form.Item>
 
@@ -107,7 +96,7 @@ export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>
                         }}
                         onBlur={handleBlur}
                         name={`links[${key}].title`}
-                        value={values.links[key].title}
+                        value={[...(values.links || [])][key].title}
                         placeholder="Название"
                       />
                       <ErrorMessage name={`links[${key}].title`} />
@@ -121,7 +110,7 @@ export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>
                         onBlur={handleBlur}
                         status={isUrlHasError ? 'error' : ''}
                         name={`links[${key}].url`}
-                        value={values.links[key].url}
+                        value={[...(values.links || [])][key].url}
                         placeholder="URL"
                       />
                       <ErrorMessage name={`links[${key}].url`} />
@@ -142,14 +131,8 @@ export const CreateGameInnerForm = (props: FormikProps<GameCreateFormValuesType>
       </Form.Item>
 
       <div className={style.editGameForm__actions}>
-        <Button onClick={() => handleSubmit()} htmlType="submit">
-          Создать и перейти к новой
-        </Button>
-        <Button onClick={() => handleSubmit()} htmlType="submit">
-          Создать и продолжить редактирование
-        </Button>
-        <Button type="primary" onClick={() => handleSubmit()} htmlType="submit">
-          Создать
+        <Button htmlType="submit" type="primary" onClick={() => handleSubmit()}>
+          Обновить
         </Button>
       </div>
     </Form>
