--- client/src/pages/GameEditPage/components/InnerForm/index.ts
@@ -1,2 +1,0 @@
-export * from './GameCreateInnerForm';
-export * from './GameUpdateInnerForm';
