+++ client/src/pages/GameEditPage/components/index.ts
@@ -1,2 +1,1 @@
-export * from './InnerForm';
 export * from './GameEditForm';
